+++ Flutter/rtc_conference_tui_kit/lib/common/widgets/index.dart
@@ -12,3 +12,4 @@ export './single_select_list.dart';
 export './toast.dart';
 export './user_info.dart';
 export './dialog.dart';
+export './volume_bar.dart';