+++ Flutter/rtc_conference_tui_kit/lib/pages/conference_main/view.dart
@@ -1,13 +1,26 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:rtc_conference_tui_kit/common/index.dart';
+import 'package:rtc_conference_tui_kit/conference/conference_params.dart';
+import 'package:rtc_conference_tui_kit/conference/conference_observer.dart';
 import 'package:rtc_conference_tui_kit/pages/conference_main/widgets/bottom_view/widgets/widgets.dart';
 import 'package:rtc_conference_tui_kit/pages/conference_main/widgets/widgets.dart';
 
 import 'index.dart';
 
 class ConferenceMainPage extends GetView<ConferenceMainController> {
-  const ConferenceMainPage({Key? key}) : super(key: key);
+  const ConferenceMainPage(
+      {this.roomId,
+      this.isCreateRoom,
+      this.conferenceParams,
+      this.conferenceObserver,
+      Key? key})
+      : super(key: key);
+
+  final String? roomId;
+  final bool? isCreateRoom;
+  final ConferenceParams? conferenceParams;
+  final ConferenceObserver? conferenceObserver;
 
   Widget _buildView() {
     return GestureDetector(
@@ -17,7 +30,7 @@ class ConferenceMainPage extends GetView<ConferenceMainController> {
       behavior: HitTestBehavior.translucent,
       child: WillPopScope(
         onWillPop: () async {
-          Get.bottomSheet(const ExitWidget());
+          showConferenceBottomSheet(const ExitWidget());
           return false;
         },
         child: Stack(
@@ -31,9 +44,11 @@ class ConferenceMainPage extends GetView<ConferenceMainController> {
                       child: SizedBox(
                         width: Get.width,
                         height: 660.0.scale375Height(),
-                        child: RoomStore.to.currentUser.hasScreenStream.value
-                            ? const LocalScreenSharingWidget()
-                            : const VideoPageTurningPage(),
+                        child: controller.isEnteredRoom.value
+                            ? RoomStore.to.currentUser.hasScreenStream.value
+                                ? const LocalScreenSharingWidget()
+                                : const VideoPageTurningPage()
+                            : const SizedBox.shrink(),
                       ),
                     ),
                     SizedBox(height: 50.0.scale375Height()),
@@ -41,16 +56,21 @@ class ConferenceMainPage extends GetView<ConferenceMainController> {
                 );
               },
             ),
-            Obx(() => Visibility(
-                visible: controller.areWidgetsVisible.value,
-                child: controller.topViewWidget)),
+            Obx(
+              () => Visibility(
+                visible: controller.areWidgetsVisible.value &&
+                    controller.isEnteredRoom.value,
+                child: const TopViewWidget(),
+              ),
+            ),
             Column(
               children: [
                 const Expanded(child: SizedBox()),
                 Obx(
                   () => Visibility(
-                    visible: controller.areWidgetsVisible.value,
-                    child: controller.bottomWidget,
+                    visible: controller.areWidgetsVisible.value &&
+                        controller.isEnteredRoom.value,
+                    child: const BottomViewWidget(),
                   ),
                 ),
                 Obx(
@@ -71,7 +91,11 @@ class ConferenceMainPage extends GetView<ConferenceMainController> {
   @override
   Widget build(BuildContext context) {
     return GetBuilder<ConferenceMainController>(
-      init: ConferenceMainController(),
+      init: ConferenceMainController(
+          roomId: roomId,
+          isCreateRoom: isCreateRoom,
+          conferenceParams: conferenceParams,
+          conferenceObserver: conferenceObserver),
       id: "conference_main",
       builder: (_) {
         return Scaffold(