+++ Flutter/rtc_conference_tui_kit/lib/pages/conference_main/widgets/user_list/widgets/user_item.dart
@@ -22,7 +22,7 @@ class UserListItem extends GetView<UserListController> {
           behavior: HitTestBehavior.opaque,
           onTap: () {
             if (controller.isAbleToControlUser(userModel)) {
-              Get.bottomSheet(UserControlWidget(userModel: userModel),
+              showConferenceBottomSheet(UserControlWidget(userModel: userModel),
                   isScrollControlled: true);
             }
           },