+++ Flutter/rtc_conference_tui_kit/lib/pages/conference_main/widgets/video_seat/video_page_turning/view.dart
@@ -17,6 +17,7 @@ class VideoPageTurningPage extends GetView<VideoPageTurningController> {
         Expanded(
           child: Obx(
             () => PageView.builder(
+              controller: controller.pageController,
               onPageChanged: (value) {
                 controller.currentIndex = value;
               },