+++ Flutter/rtc_conference_tui_kit/lib/rtc_conference_tui_kit.dart
@@ -1,0 +1,8 @@
+library rtc_conference_tui_kit;
+
+export './conference/conference_session.dart';
+export './conference/tui_room_kit.dart';
+export './conference/conference_params.dart';
+export './conference/conference_observer.dart';
+export './conference/conference_error.dart';
+export './pages/conference_main/view.dart';